Fixes Already
I was going through the app to test for errors and found quite a few dumb ones in the Federation and Klingon character creation forms. I'll be running through the Romulan form tomorrow night.

Consider the application without a version number to be 1.0. With that, I give you version 1.1.

BTW, GitHub gives me the option to create a "new release," but in doing so doesn't give me the option to upload the file itself. That tells me that "new release" must create a new repository or something where I later add the file. Or maybe not. So, if this is the least efficient way of doing this, oops.