Version 1.3
Unfortunately, the limitations of Access have reared their ugly heads again. Version 1.3 is now 26M even in its bear-bones form. So, I had to put it in a zip file. I hope that doesn't create issues for you. The application already restricts too many of you from using it as it is. Version 1.3 adds NPC generation for all races, and "detailed enlisted" generation for Federation characters. The latter has the capacity to produce a report, which can be saved to PDF as normal, but you can't save the data yet. I'll get to that shortly. I also plan generation of non-sentient beings and solar systems. Then I'll be focused on making web applications to replace these.