Version 1.7
This is the final version of the application as far as features are concerned. Going forward, all my focus will go to developing the web application. If anyone finds any bugs, those will be corrected. One thing that I'll definitely work on is dealing with a silly restriction Access has; to-wit: Including commas, apostrophes, and quotation marks in your text can often result in corrupted data. I can't stop users from doing that, so I need to find a way to work around it. Also, if any of you find any other genuine bugs, I'll deal with those. But new features? That will have to wait for the web application. As far as what those features are, I believe the only ones missing are creating starships and constructing starship battle groups. This will never be a true campaign manager, so I believe those are the only two features missing. Of course, you can still get them in my 2012 application if you can stomach the user interface. :-)