Version 1.9
That didn't take long. I created a routine that replaces single and couple quotes with backward tic marks (`) and commas with a space and a dash ( -) in a given text field. This routine will been applied whenever a user leaves the field. For example, if you type XYZ'XYZ"XYZ, XYZ as your character's name, as soon as you move away from that field, the name will be changed to XYZ`XYZ`XYZ - XYZ. Unless I missed one, this will occur whenever you're naming a character, world, solar system, game, art form, etc. It's not particularly elegant, but it allows the user to proceed with all the functions of the application.